refactor(llm): one subject per file, and one answer per question
The package answered several questions more than once. Four functions built
the same "vendor:auth_method" identity string. Two classifiers read the same
failure, each carrying its own retryable and context-exceeded wrapper. Every
registry function existed twice — once as a method nothing outside the package
called — and one of those pairs took a read lock recursively, which deadlocks
the moment a registration is waiting between the two. Files were named for the
package (llm.go, types.go) rather than for what is in them.

- llmerr folds into errors.go, merged with vendor_errors.go. The SDK already
  classifies a failure from the provider's own typed error, and its table is
  the better one: it excludes the throttles that are worded like an overflow,
  which llmerr's did not, so a Bedrock rate limit was being told to compact.
  What stays here is the translation onto core.RetryableError and
  core.ContextExceededError. anthropic-sdk-go and openai-go are no longer
  direct dependencies.
- providerKey replaces makeProviderKey, makemodelCacheKey, providerName and
  Meta.Key. The registry's duplicate methods go with them, and the status scan
  snapshots rather than classifying under the lock.
- Deleted: SanitizeToolMessages (the SDK repairs history now), Client.send and
  four FakeLLM methods (each called only by its own test), the empty Options
  struct, and CATALOGS.md, which documents packages that no longer exist.
- FakeLLM implements Provider, so the adapter wrapping it in the integration
  helpers is gone.
- Files renamed for their subject: client.go, provider.go, conn.go,
  modelcache.go, effort.go, cost.go, fake.go. Name becomes ProviderID and Info
  becomes AuthMethodStatus — a slug is not a name, and neither was "Info".
