+++ internal/llm/cost.go
@@ -1,39 +1,104 @@
+// What a turn costs, and who can say.
+//
+// Two things live here because they are one subject: an amount of money and
+// the per-vendor pricing that produces one. Currencies are kept apart rather
+// than summed — there is no exchange rate here, and adding across them would
+// invent a number.
 package llm
 
+import (
+	"slices"
+	"strings"
+)
+
 // CostEstimator computes the money cost of a single inference from its token
-// usage, returning (_, false) when the model's pricing is unknown. Each
-// provider package registers its own in init() via RegisterCostEstimator, so
-// callers price a turn through EstimateCost without importing provider
-// internals or switching on the provider name.
+// usage, returning (_, false) when the model's pricing is unknown. The vendor
+// table registers one per provider, so callers price a turn through
+// EstimateCost without knowing which vendor served it.
 type CostEstimator func(modelID string, usage Usage) (Money, bool)
 
-// RegisterCostEstimator registers a provider's cost estimator. Call once per
-// provider package, typically alongside Register() in init().
-func RegisterCostEstimator(provider Name, fn CostEstimator) {
-	globalRegistry.RegisterCostEstimator(provider, fn)
+// RegisterCostEstimator records a provider's pricing.
+func RegisterCostEstimator(provider ProviderID, estimate CostEstimator) {
+	globalRegistry.mu.Lock()
+	defer globalRegistry.mu.Unlock()
+	globalRegistry.costs[provider] = estimate
 }
 
-// RegisterCostEstimator registers a provider's cost estimator.
-func (r *Registry) RegisterCostEstimator(provider Name, fn CostEstimator) {
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	r.costEstimators[provider] = fn
-}
+// EstimateCost prices one turn, reporting ok=false when the provider has no
+// registered pricing or the model has no published rate — which San renders as
+// "--" rather than as free.
+func EstimateCost(provider ProviderID, modelID string, usage Usage) (Money, bool) {
+	globalRegistry.mu.RLock()
+	estimate, ok := globalRegistry.costs[provider]
+	globalRegistry.mu.RUnlock()
 
-// EstimateCost dispatches to the estimator registered for provider, or returns
-// (Money{}, false) when the provider has no registered pricing.
-func EstimateCost(provider Name, modelID string, usage Usage) (Money, bool) {
-	return globalRegistry.EstimateCost(provider, modelID, usage)
-}
-
-// EstimateCost dispatches to the estimator registered for provider, or returns
-// (Money{}, false) when the provider has no registered pricing.
-func (r *Registry) EstimateCost(provider Name, modelID string, usage Usage) (Money, bool) {
-	r.mu.RLock()
-	fn, ok := r.costEstimators[provider]
-	r.mu.RUnlock()
 	if !ok {
 		return Money{}, false
 	}
-	return fn(modelID, usage)
+	return estimate(modelID, usage)
+}
+
+type Currency string
+
+const (
+	CurrencyCNY Currency = "CNY"
+	CurrencyUSD Currency = "USD"
+)
+
+// Money is one amount in one currency — what a single provider charges for a
+// single call. Amounts are not summed across currencies; see CostTotal.
+type Money struct {
+	Amount   float64
+	Currency Currency
+}
+
+func (m Money) IsZero() bool {
+	return m.Amount == 0 || m.Currency == ""
+}
+
+// CostTotal accumulates spend that may span currencies. A session can switch
+// providers, and providers do not agree on one: MiniMax prices in CNY while
+// DeepSeek, MiMo and Ollama price in USD. There is no exchange rate here, so
+// adding across them would invent a number — each currency is kept on its own
+// and rendered side by side.
+//
+// The zero value is an empty total. Add returns a new CostTotal rather than
+// mutating, so the value can be copied freely; env is passed around by value.
+type CostTotal struct {
+	// entries holds at most one Money per currency, sorted by currency code so
+	// the status bar renders them in a stable order rather than reshuffling as
+	// the map iteration order changes.
+	entries []Money
+}
+
+// NewCostTotal starts a total from a single amount.
+func NewCostTotal(m Money) CostTotal {
+	return CostTotal{}.Add(m)
+}
+
+// Add folds m into the total, keeping it on its own currency's running sum.
+func (c CostTotal) Add(m Money) CostTotal {
+	if m.IsZero() {
+		return c
+	}
+	entries := slices.Clone(c.entries)
+	if i := slices.IndexFunc(entries, func(e Money) bool { return e.Currency == m.Currency }); i >= 0 {
+		entries[i].Amount += m.Amount
+		return CostTotal{entries: entries}
+	}
+	entries = append(entries, m)
+	slices.SortFunc(entries, func(a, b Money) int {
+		return strings.Compare(string(a.Currency), string(b.Currency))
+	})
+	return CostTotal{entries: entries}
+}
+
+// IsZero reports whether nothing has been spent.
+func (c CostTotal) IsZero() bool {
+	return len(c.entries) == 0
+}
+
+// Amounts returns the per-currency totals, ordered by currency code.
+func (c CostTotal) Amounts() []Money {
+	return slices.Clone(c.entries)
 }
