+++ internal/llm/registry.go
@@ -3,230 +3,112 @@ package llm
 import (
 	"context"
 	"fmt"
-	"sort"
+	"maps"
+	"slices"
 	"strings"
 	"sync"
 
 	"github.com/genai-io/san/internal/secret"
 )
 
-// registryEntry holds a provider's metadata and factory
+// The table of everything San can connect to.
+//
+// There is one of it, filled at init by the vendor table and read from the
+// picker, the status bar and every entry point that opens a connection. It is
+// a package-level singleton rather than a value callers pass around, because a
+// second one would be a second answer to "which providers exist".
+
+// registry holds what is known about each provider/auth-method pair. The four
+// maps are keyed and locked together so a registration is visible whole.
+type registry struct {
+	mu             sync.RWMutex
+	entries        map[string]registryEntry       // key: providerKey
+	displays       map[ProviderID]ProviderDisplay // provider-level UI presentation
+	costs          map[ProviderID]CostEstimator   // per-provider turn-cost pricing
+	authenticators map[string]Authenticator       // key: providerKey; interactive (OAuth) login
+}
+
+// registryEntry is one provider/auth-method pair: what it is, and how to open it.
 type registryEntry struct {
 	meta    Meta
 	factory Factory
 }
 
-// Registry manages provider registration and discovery
-type Registry struct {
-	mu              sync.RWMutex
-	entries         map[string]registryEntry // key: "provider:authMethod"
-	providerDisplay map[Name]ProviderDisplay // provider-level UI presentation
-	costEstimators  map[Name]CostEstimator   // per-provider turn-cost pricing
-	authenticators  map[string]Authenticator // key: "provider:authMethod"; interactive (OAuth) login
+var globalRegistry = &registry{
+	entries:        make(map[string]registryEntry),
+	displays:       make(map[ProviderID]ProviderDisplay),
+	costs:          make(map[ProviderID]CostEstimator),
+	authenticators: make(map[string]Authenticator),
 }
 
-// globalRegistry is the default registry instance
-var globalRegistry = &Registry{
-	entries:         make(map[string]registryEntry),
-	providerDisplay: make(map[Name]ProviderDisplay),
-	costEstimators:  make(map[Name]CostEstimator),
-	authenticators:  make(map[string]Authenticator),
-}
-
-// Register registers a provider with its metadata and factory
-func Register(meta Meta, factory Factory) {
-	globalRegistry.Register(meta, factory)
-}
-
-// RegisterProviderDisplay registers a provider's UI presentation (display name and order).
-// Call once per provider package — typically alongside the first Register() call.
-func RegisterProviderDisplay(provider Name, pd ProviderDisplay) {
-	globalRegistry.RegisterProviderDisplay(provider, pd)
-}
-
-// Unregister removes a provider/auth-method entry from the global registry.
-func Unregister(provider Name, authMethod AuthMethod) {
-	globalRegistry.Unregister(provider, authMethod)
-}
-
-// Register registers a provider with its metadata and factory
-func (r *Registry) Register(meta Meta, factory Factory) {
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	r.entries[meta.Key()] = registryEntry{
-		meta:    meta,
-		factory: factory,
-	}
-}
-
-// RegisterProviderDisplay registers a provider's UI presentation.
-func (r *Registry) RegisterProviderDisplay(provider Name, pd ProviderDisplay) {
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	r.providerDisplay[provider] = pd
-}
-
-// Unregister removes a provider/auth-method entry from the registry.
-func (r *Registry) Unregister(provider Name, authMethod AuthMethod) {
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	delete(r.entries, makeProviderKey(provider, authMethod))
-}
-
-// GetProvider returns a provider instance for the given provider and auth method
-func GetProvider(ctx context.Context, provider Name, authMethod AuthMethod) (Provider, error) {
-	return globalRegistry.GetProvider(ctx, provider, authMethod)
-}
-
-// GetProvider returns a provider instance for the given provider and auth method
-func (r *Registry) GetProvider(ctx context.Context, provider Name, authMethod AuthMethod) (Provider, error) {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	entry, ok := r.entries[makeProviderKey(provider, authMethod)]
-	if !ok {
-		return nil, fmt.Errorf("provider not registered: %s:%s", provider, authMethod)
-	}
-
-	return entry.factory(ctx)
-}
-
-// GetMeta returns the metadata for a specific provider configuration
-func GetMeta(provider Name, authMethod AuthMethod) (Meta, bool) {
-	return globalRegistry.GetMeta(provider, authMethod)
-}
-
-// GetMeta returns the metadata for a specific provider configuration
-func (r *Registry) GetMeta(provider Name, authMethod AuthMethod) (Meta, bool) {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	entry, ok := r.entries[makeProviderKey(provider, authMethod)]
-	if !ok {
-		return Meta{}, false
-	}
-	return entry.meta, true
-}
-
-// makeProviderKey creates a unique key for provider and auth method combination
-func makeProviderKey(provider Name, authMethod AuthMethod) string {
+// providerKey identifies one provider/auth-method pair, as "vendor:auth_method".
+//
+// It is the form the registry keys entries by, the store keys cached model
+// listings by, and a provider reports as its own Name. One function so the
+// three can never drift into three slightly different strings.
+func providerKey(provider ProviderID, authMethod AuthMethod) string {
 	return string(provider) + ":" + string(authMethod)
 }
 
-// IsReady checks if all required environment variables are set for a provider
-func IsReady(meta Meta) bool {
-	return globalRegistry.IsReady(meta)
+// ---------------------------------------------------------------------------
+// Registration
+// ---------------------------------------------------------------------------
+
+// Register records a provider auth method and the factory that opens it.
+func Register(meta Meta, factory Factory) {
+	globalRegistry.mu.Lock()
+	defer globalRegistry.mu.Unlock()
+	globalRegistry.entries[providerKey(meta.Provider, meta.AuthMethod)] = registryEntry{meta: meta, factory: factory}
 }
 
-// IsReady checks if all required environment variables are set for a provider
-func (r *Registry) IsReady(meta Meta) bool {
-	for _, envVar := range meta.EnvVars {
-		if secret.Resolve(envVar) == "" {
-			return false
-		}
-	}
-	// No env vars means an OAuth/credential-based auth method (e.g. ChatGPT
-	// Subscription). Without stored credentials it is not ready — treat it as
-	// "not configured" rather than showing a misleading yellow "Available" icon.
-	if len(meta.EnvVars) == 0 {
-		if a := r.authenticator(meta.Provider, meta.AuthMethod); a != nil {
-			if sca, ok := a.(StoredCredentialAuthenticator); ok {
-				return sca.HasCredentials()
-			}
-		}
-		return false
-	}
-	return true
+// RegisterProviderDisplay records a provider's UI presentation (display name
+// and order), which is shared by all of its auth methods.
+func RegisterProviderDisplay(provider ProviderID, display ProviderDisplay) {
+	globalRegistry.mu.Lock()
+	defer globalRegistry.mu.Unlock()
+	globalRegistry.displays[provider] = display
 }
 
-// Status represents the connection status of a provider
-type Status string
-
-const (
-	StatusConnected     Status = "connected"
-	StatusAvailable     Status = "available"
-	StatusNotConfigured Status = "not_configured"
-)
-
-// Info contains provider metadata with its current status
-type Info struct {
-	Meta   Meta
-	Status Status
+// Unregister removes a provider/auth-method entry.
+func Unregister(provider ProviderID, authMethod AuthMethod) {
+	globalRegistry.mu.Lock()
+	defer globalRegistry.mu.Unlock()
+	delete(globalRegistry.entries, providerKey(provider, authMethod))
 }
 
-// GetProvidersWithStatus returns all providers grouped by provider name with their status
-func GetProvidersWithStatus(store *Store) map[Name][]Info {
-	return globalRegistry.GetProvidersWithStatus(store)
+// ---------------------------------------------------------------------------
+// Lookup
+// ---------------------------------------------------------------------------
+
+// GetProvider opens a connection to a registered provider auth method.
+func GetProvider(ctx context.Context, provider ProviderID, authMethod AuthMethod) (Provider, error) {
+	globalRegistry.mu.RLock()
+	entry, ok := globalRegistry.entries[providerKey(provider, authMethod)]
+	globalRegistry.mu.RUnlock()
+
+	if !ok {
+		return nil, fmt.Errorf("provider not registered: %s:%s", provider, authMethod)
+	}
+	return entry.factory(ctx)
 }
 
-// GetProvidersWithStatus returns all providers grouped by provider name with their status
-func (r *Registry) GetProvidersWithStatus(store *Store) map[Name][]Info {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
+// GetMeta returns the metadata registered for a provider auth method.
+func GetMeta(provider ProviderID, authMethod AuthMethod) (Meta, bool) {
+	globalRegistry.mu.RLock()
+	defer globalRegistry.mu.RUnlock()
 
-	result := make(map[Name][]Info)
-
-	for _, entry := range r.entries {
-		var status Status
-		if store.IsConnected(entry.meta.Provider, entry.meta.AuthMethod) {
-			status = StatusConnected
-		} else if r.IsReady(entry.meta) {
-			status = StatusAvailable
-		} else {
-			status = StatusNotConfigured
-		}
-
-		info := Info{
-			Meta:   entry.meta,
-			Status: status,
-		}
-		result[entry.meta.Provider] = append(result[entry.meta.Provider], info)
-	}
-
-	return result
-}
-
-// ProvidersByOrder returns unique provider names sorted by their Order field.
-func ProvidersByOrder() []Name {
-	return globalRegistry.ProvidersByOrder()
-}
-
-// ProvidersByOrder returns unique provider names sorted by their Order field.
-func (r *Registry) ProvidersByOrder() []Name {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	type providerOrder struct {
-		name  Name
-		order int
-	}
-	providers := make([]providerOrder, 0, len(r.providerDisplay))
-	for name, pd := range r.providerDisplay {
-		providers = append(providers, providerOrder{name, pd.Order})
-	}
-	sort.Slice(providers, func(i, j int) bool {
-		return providers[i].order < providers[j].order
-	})
-	result := make([]Name, len(providers))
-	for i, p := range providers {
-		result[i] = p.name
-	}
-	return result
+	entry, ok := globalRegistry.entries[providerKey(provider, authMethod)]
+	return entry.meta, ok
 }
 
 // IsProvider reports whether name is a registered provider vendor. It is the
 // source of truth for telling a "vendor/model" routing ref apart from a bare
 // model id that merely contains a slash.
-func IsProvider(name Name) bool {
-	return globalRegistry.IsProvider(name)
-}
+func IsProvider(name ProviderID) bool {
+	globalRegistry.mu.RLock()
+	defer globalRegistry.mu.RUnlock()
 
-// IsProvider reports whether name is a registered provider vendor.
-func (r *Registry) IsProvider(name Name) bool {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	_, ok := r.providerDisplay[name]
+	_, ok := globalRegistry.displays[name]
 	return ok
 }
 
@@ -236,26 +118,101 @@ func (r *Registry) IsProvider(name Name) bool {
 // contains a slash (e.g. mimo's "xiaomi/mimo-v2-flash") is not, so it reports
 // ok=false and the caller keeps the ref on the current provider. A known but
 // unconnected vendor still parses here; resolving it is the caller's job.
-func ParseVendorModel(ref string) (vendor Name, model string, ok bool) {
+func ParseVendorModel(ref string) (vendor ProviderID, model string, ok bool) {
 	v, m, found := strings.Cut(ref, "/")
-	if !found || v == "" || m == "" || !IsProvider(Name(v)) {
+	if !found || v == "" || m == "" || !IsProvider(ProviderID(v)) {
 		return "", "", false
 	}
-	return Name(v), m, true
+	return ProviderID(v), m, true
 }
 
-// ProviderDisplayName returns the provider-level display name for a provider.
-func ProviderDisplayName(p Name) string {
-	return globalRegistry.ProviderDisplayName(p)
-}
+// ProviderDisplayName returns a provider's human-readable name, falling back
+// to its identifier when none was registered.
+func ProviderDisplayName(provider ProviderID) string {
+	globalRegistry.mu.RLock()
+	defer globalRegistry.mu.RUnlock()
 
-// ProviderDisplayName returns the provider-level display name for a provider.
-func (r *Registry) ProviderDisplayName(p Name) string {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	if pd, ok := r.providerDisplay[p]; ok {
-		return pd.Name
+	if display, ok := globalRegistry.displays[provider]; ok {
+		return display.Name
 	}
-	return string(p)
+	return string(provider)
+}
+
+// ProvidersByOrder returns every registered provider, in display order.
+func ProvidersByOrder() []ProviderID {
+	globalRegistry.mu.RLock()
+	displays := maps.Clone(globalRegistry.displays)
+	globalRegistry.mu.RUnlock()
+
+	names := slices.Collect(maps.Keys(displays))
+	slices.SortFunc(names, func(a, b ProviderID) int {
+		if order := displays[a].Order - displays[b].Order; order != 0 {
+			return order
+		}
+		// Equal orders would otherwise come back in map order, which Go
+		// randomizes — the picker would reshuffle between renders.
+		return strings.Compare(string(a), string(b))
+	})
+	return names
+}
+
+// ---------------------------------------------------------------------------
+// Connection status
+// ---------------------------------------------------------------------------
+
+// Status is how far a provider auth method is from being usable.
+type Status string
+
+const (
+	StatusConnected     Status = "connected"
+	StatusAvailable     Status = "available"
+	StatusNotConfigured Status = "not_configured"
+)
+
+// AuthMethodStatus is one registered auth method with its current status.
+type AuthMethodStatus struct {
+	Meta   Meta
+	Status Status
+}
+
+// GetProvidersWithStatus reports every registered auth method, grouped by
+// provider, with how far each is from being usable.
+func GetProvidersWithStatus(store *Store) map[ProviderID][]AuthMethodStatus {
+	// Snapshot under the lock and classify outside it: IsReady reaches into
+	// the secret store and the authenticators, and holding a read lock across
+	// that would block every registration behind it.
+	globalRegistry.mu.RLock()
+	entries := slices.Collect(maps.Values(globalRegistry.entries))
+	globalRegistry.mu.RUnlock()
+
+	result := make(map[ProviderID][]AuthMethodStatus, len(entries))
+	for _, entry := range entries {
+		status := StatusNotConfigured
+		switch {
+		case store.IsConnected(entry.meta.Provider, entry.meta.AuthMethod):
+			status = StatusConnected
+		case IsReady(entry.meta):
+			status = StatusAvailable
+		}
+		result[entry.meta.Provider] = append(result[entry.meta.Provider], AuthMethodStatus{Meta: entry.meta, Status: status})
+	}
+	return result
+}
+
+// IsReady reports whether an auth method's credentials are in place, so it can
+// be shown as available to connect.
+func IsReady(meta Meta) bool {
+	// No env vars means the credential comes from an interactive sign-in (e.g.
+	// a ChatGPT subscription). Without stored credentials it is not ready —
+	// treat it as "not configured" rather than showing a misleading
+	// "Available".
+	if len(meta.EnvVars) == 0 {
+		return HasInteractiveCredentials(meta.Provider, meta.AuthMethod)
+	}
+	for _, envVar := range meta.EnvVars {
+		if secret.Resolve(envVar) == "" {
+			return false
+		}
+	}
+	return true
 }
