+++ tests/integration/testutil/helpers.go
@@ -15,10 +15,10 @@ import (
 // Client helpers
 // ---------------------------------------------------------------------------
 
-// NewTestClient wraps a FakeLLM in a llm.Client ready for use in loops
-// or compact calls. This avoids repeating the FakeProvider wiring in every test.
+// NewTestClient fixes a model on a FakeLLM, ready for use in loops or compact
+// calls. llm.FakeLLM is a llm.Provider, so there is nothing to adapt.
 func NewTestClient(fake *llm.FakeLLM) *llm.Client {
-	return llm.NewClient(&FakeProvider{Client: fake}, "fake-model", 8192)
+	return llm.NewClient(fake, "fake-model", 8192)
 }
 
 // ---------------------------------------------------------------------------
@@ -103,22 +103,10 @@ func (f *fakeTool) Execute(_ context.Context, _ map[string]any, _ string) toolre
 // Fake / mock providers
 // ---------------------------------------------------------------------------
 
-// FakeProvider wraps a FakeClient as a llm.Provider.
-// Use this when the code under test expects a llm.Provider and you
-// want to control responses via FakeClient.
-type FakeProvider struct {
-	Client *llm.FakeLLM
-}
-
-func (p *FakeProvider) Stream(ctx context.Context, opts llm.CompletionOptions) <-chan llm.StreamChunk {
-	return p.Client.Stream(ctx, opts.Messages, opts.Tools, opts.SystemPrompt)
-}
-func (p *FakeProvider) ListModels(_ context.Context) ([]llm.ModelInfo, error) { return nil, nil }
-func (p *FakeProvider) Name() string                                          { return p.Client.Name() }
-
-// MockProvider is a standalone llm.Provider backed by a response queue.
-// Unlike FakeProvider, it does not require a FakeClient — use this when the
-// code under test (e.g., agent.Executor) creates its own client internally.
+// MockProvider is a standalone llm.Provider backed by a response queue. Unlike
+// llm.FakeLLM it records nothing — use it when the code under test (e.g.
+// agent.Executor) builds its own client internally and the test only needs a
+// provider that answers.
 type MockProvider struct {
 	Responses []llm.CompletionResponse
 	callIdx   int
